# openapi/transcendence.v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Transcendence API",
    "description": "League analytics API with app API-key and user JWT authentication.",
    "version": "v1"
  },
  "paths": {
...
    "/api/analytics/cache/invalidate": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          { }
        ]
      }
    },
...
    "/api/analytics/champions/{championId}/winrates": {
      "get": {
        "tags": [
          "ChampionAnalytics"
        ],
        "parameters": [
          {
            "name": "championId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "rankTier",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionWinRateSummary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionWinRateSummary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionWinRateSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/champions/{championId}/builds": {
      "get": {
        "tags": [
          "ChampionAnalytics"
        ],
        "parameters": [
          {
            "name": "championId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rankTier",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionBuildsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionBuildsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionBuildsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/champions/{championId}/pro-builds": {
      "get": {
        "tags": [
          "ChampionAnalytics"
        ],
        "parameters": [
          {
            "name": "championId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "region",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "patch",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionProBuildsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionProBuildsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionProBuildsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/champions/{championId}/matchups": {
      "get": {
        "tags": [
          "ChampionAnalytics"
        ],
        "parameters": [
          {
            "name": "championId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rankTier",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionMatchupsResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionMatchupsResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChampionMatchupsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/analytics/champions/cache/invalidate": {
      "post": {
        "tags": [
          "ChampionAnalytics"
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          { }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChampionBuildDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "coreItems": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "situationalItems": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "primaryStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "subStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "primaryRunes": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "subRunes": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "statShards": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "games": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ChampionBuildsResponse": {
        "type": "object",
        "properties": {
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "rankTier": {
            "type": "string",
            "nullable": true
          },
          "patch": {
            "type": "string",
            "nullable": true
          },
          "globalCoreItems": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "builds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChampionBuildDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChampionMatchupsResponse": {
        "type": "object",
        "properties": {
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "rankTier": {
            "type": "string",
            "nullable": true
          },
          "patch": {
            "type": "string",
            "nullable": true
          },
          "counters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchupEntryDto"
            },
            "nullable": true
          },
          "favorableMatchups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchupEntryDto"
            },
            "nullable": true
          },
          "allMatchups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchupEntryDto"
            },
            "nullable": true
          },
          "timelineCoverageRatio": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "timelineSampleSize": {
            "type": "integer",
            "format": "int32"
          },
          "timelineDataFreshnessUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChampionProBuildsResponse": {
        "type": "object",
        "properties": {
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "patch": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "string",
            "nullable": true
          },
          "recentProMatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProMatchBuildDto"
            },
            "nullable": true
          },
          "topPlayers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProPlayerSummaryDto"
            },
            "nullable": true
          },
          "commonBuilds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommonProBuildDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ChampionWinRateDto": {
        "type": "object",
        "properties": {
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "rankTier": {
            "type": "string",
            "nullable": true
          },
          "games": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          },
          "pickRate": {
            "type": "number",
            "format": "double"
          },
          "banRate": {
            "type": "number",
            "format": "double"
          },
          "roleRank": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rolePopulation": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "patch": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ChampionWinRateSummary": {
        "type": "object",
        "properties": {
          "championId": {
            "type": "integer",
            "format": "int32"
          },
          "patch": {
            "type": "string",
            "nullable": true
          },
          "byRoleTier": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChampionWinRateDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CommonProBuildDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "games": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "MatchupEntryDto": {
        "type": "object",
        "properties": {
          "opponentChampionId": {
            "type": "integer",
            "format": "int32"
          },
          "games": {
            "type": "integer",
            "format": "int32"
          },
          "wins": {
            "type": "integer",
            "format": "int32"
          },
          "losses": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          },
          "avgGoldDiffAt15": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "avgXpDiffAt15": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProMatchBuildDto": {
        "type": "object",
        "properties": {
          "matchId": {
            "type": "string",
            "nullable": true
          },
          "playerName": {
            "type": "string",
            "nullable": true
          },
          "teamName": {
            "type": "string",
            "nullable": true
          },
          "win": {
            "type": "boolean"
          },
          "playedAt": {
            "type": "integer",
            "format": "int64"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "primaryStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "subStyleId": {
            "type": "integer",
            "format": "int32"
          },
          "primaryRunes": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "subRunes": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "statShards": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProPlayerSummaryDto": {
        "type": "object",
        "properties": {
          "playerName": {
            "type": "string",
            "nullable": true
          },
          "teamName": {
            "type": "string",
            "nullable": true
          },
          "games": {
            "type": "integer",
            "format": "int32"
          },
          "winRate": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "App authentication key for operational endpoints.",
        "name": "X-API-Key",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "User JWT access token from /api/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}